Add Vector Expression node with separate axes and rename old vector node
